@@ -1,3 +1,3 @@
 module github.com/thevilledev/go-thespine
 
-go 1.20
+go 1.23.2
